+++ pkg/data/domain/cluster/common.go
@@ -12,8 +12,8 @@ func (s *Service) Get(ctx context.Context, options GetOptions) (Resource, error)
 	var resource Resource
 	var err error
 
-	if len(options.ID) > 0 {
-		resource, err = s.getById(ctx, options.Provider, options.ID, options.Namespace)
+	if len(options.Name) > 0 {
+		resource, err = s.getByName(ctx, options.Provider, options.Name, options.Namespace)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -27,20 +27,20 @@ func (s *Service) Get(ctx context.Context, options GetOptions) (Resource, error)
 	return resource, nil
 }
 
-func (s *Service) getById(ctx context.Context, provider, id, namespace string) (Resource, error) {
+func (s *Service) getByName(ctx context.Context, provider, name, namespace string) (Resource, error) {
 	var err error
 
 	var cluster Resource
 	{
 		switch provider {
 		case key.ProviderAWS:
-			cluster, err = s.getByIdAWS(ctx, id, namespace)
+			cluster, err = s.getByNameAWS(ctx, name, namespace)
 			if err != nil {
 				return nil, microerror.Mask(err)
 			}
 
 		case key.ProviderAzure:
-			cluster, err = s.getByIdAzure(ctx, id, namespace)
+			cluster, err = s.getByNameAzure(ctx, name, namespace)
 			if err != nil {
 				return nil, microerror.Mask(err)
 			}