+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+- `template cluster`:
+  - Deprecated the `--cluster-id` flag.
+  - Breaking: the `--name` flag changed purpose to set the cluster's unique identifier.
+  - The `--description` flag has been added to set the user-friendly description.
 - `template nodepool`:
   - Deprecated the `--cluster-id` flag, added the `--cluster-name` flag as a replacement.
   - Deprecated the `-nodepool-name` flag, add the `--description` flag as a replacement.