+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Changed
 
+- `get clusters`:
+  - The output table header `ID` has been renamed to `NAME`.
+- `get nodepools`:
+  - The `--cluster-id` flag is now deprecated, replaced with `--cluster-name`.
+  - Output column headers have been renamed from `ID` to `NAME` and from `CLUSTER ID` to `CLUSTER NAME`.
 - `template cluster`:
   - Deprecated the `--cluster-id` flag.
   - Breaking: the `--name` flag changed purpose to set the cluster's unique identifier.