+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [1.34.0] - 2021-07-30
+
 ### Added
 
 - Add `--namespace-annotations` and `--namespace-labels` flags to `template app` command to allow users to
@@ -417,7 +419,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.33.0...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.34.0...HEAD
+[1.34.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.33.0...v1.34.0
 [1.33.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.32.0...v1.33.0
 [1.32.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.31.0...v1.32.0
 [1.31.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.30.0...v1.31.0